% Generated by roxygen2 (4.0.2): do not edit by hand
\name{area_intervals}
\alias{area_intervals}
\title{Calculates the area under a curve between sets of intervals}
\usage{
area_intervals(df, beg_times, end_times, time_col = "time",
  value_col = "value", group_by = NULL, diff_time = FALSE)
}
\arguments{
\item{beg_times}{A vector of times in unix time}

\item{end_times}{A vector of times in unix time}

\item{time_col}{The name or index of the time_col column}

\item{group_by}{An index or string of the columns that serve as groups}

\item{diff_time}{A boolean determining if the givin time column is the actual timestamp, or the difference in time}
}
\description{
This function takes a data.frame/data.table, and finds the area under the curve of each interval given.
}
\details{
This function enables the area of a curve to be found at multiple intervals. By default, it is using roll_ends = true. So, if there
is a start or end time that is before or after the actual endpoints, the area will be greater/less than what you might expect. Modifications
to allow this to be more robust will be made in the future.
}
\examples{
dt <- data.table(time = 101:110, value = 1:10, group = c('a','b'))
area_intervals(dt, c(101.5, 105, 109), c(110,110,110), group_by = "group")
}

