% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_area}
\alias{calc_area}
\title{Calculate Area Under the Curve}
\usage{
calc_area(time_date, value, unit_conversion = 1, as_vector = FALSE,
  diff_time = FALSE, neg_area = FALSE, na.replace = vreplace, ...)
}
\arguments{
\item{time_date}{A numeric or date.time vector}

\item{value}{A numeric vector}

\item{unit_conversion}{The ratio of your variable to time. Be cautious of this unless inaccurate results appease you :)}

\item{as_vector}{Boolean describing if return value should be the totaled sum or a vector representing area at each `time` value}

\item{diff_time}{Boolean describing if `time` is the difference in area, or if it needs to be calculated}

\item{neg_area}{Boolean describing if negative area should be measured or it all area is positive}

\item{na.replace}{Function that replaces NA values from `value` with Last Observation Carried Forward. See \code{\link{vreplace}}}
}
\value{
The total area under the curve of `value` with respect to the time measured at `time`
}
\description{
This function takes two vectors, a time-date vector and a numeric vector and performs computations to measure to total area
under the curve. Its apprach is to use Reimann Left Sums is and aimed towards time-series data
}
\examples{
time <- 1:10
 value <- sin(1:10)

 calc_area(time, value)
 calc_area(time, value, TRUE)
 calc_area(time, value, diff_time = TRUE)
 calc_area(time, value, neg_area = TRUE)
 calc_area(time, value, TRUE, TRUE)
}

