% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand.grid.df}
\alias{expand.grid.df}
\title{Expands a list of data.frames, like expand.grid}
\usage{
expand.grid.df(...)
}
\arguments{
\item{...}{Takes a series of data.frames}
}
\description{
Expands a list of data.frames, like expand.grid
}
\examples{
df1 = data.frame(1:10, letters[1:10])
df2 = data.frame(101:110)
expand.grid.df(df1,df2)
expand.grid.df(df1,data.frame(20:22))
}

