% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.empty}
\alias{is.empty}
\title{See if value is empty}
\usage{
is.empty(x)
}
\arguments{
\item{x}{An R object}
}
\description{
This function simply checks to see if an item contains anything. Specifically, it checks
if it is null, has rows, or columns.
}
\details{
This is not meant for classes like lm. Only for plain old everyday boring R objects
}
\examples{
is.empty(1:5)
is.empty(c(1:9,3,3))
is.empty(NULL)
}

