% Generated by roxygen2 (4.0.2): do not edit by hand
\name{manipulate_replace}
\alias{manipulate_replace}
\title{Replace Values in a vector/data.frame/data.table}
\usage{
manipulate_replace(x, replace = NA, replacement = "LOCF", first = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{replace}{Value to be replaced; NA by default}

\item{replacement}{The replacement value, which is Last Observation Carried Forward by default}

\item{first}{If the first value is unknown and LOCF, will fill with first known value unless specified here}
}
\value{
A vector where all values equal to \code{replace} in \code{x} are replaced with \code{replacement}.
}
\description{
Replace NAs or a specific value in a numeric vector/data.frame/data.table
}
\examples{
tmp <- c(NA,2,3,NA,5,6)
manipulate_replace(tmp)
manipulate_replace(tmp, replace = 2)
manipulate_replace(tmp, replacement = 4)
manipulate_replace(tmp, replace = 3, replacement = 3)
manipulate_replace(tmp, first = 1)
}

