% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot functions for a time series}
\usage{
plot_timeseries(ts_obj, prd = NULL)
}
\arguments{
\item{ts_obj}{A time series, must be a \code{TimeSeries} class.}

\item{prd}{Prediction of a time series (optional), must be a \code{TimeSeries} class.}
}
\value{
The return value is a plot of the time series.
}
\description{
The function provides plots for a given time series.
}
\examples{
ar_time_series <- AR(start_values = c(1, 1), ar_params = c(0.6, 0.6), sd = 1, n = 100)
plot_timeseries(ar_time_series)

}
