% Generated by roxygen2 (4.0.2): do not edit by hand
\name{point_reduce}
\alias{dtreduce}
\alias{point_reduce}
\title{Reduce the number of points needed to plot time series data}
\usage{
point_reduce(x, tolerance)

dtreduce(dt, column, tolerance = 0.01)
}
\arguments{
\item{x}{A numeric vector}

\item{tolerance}{A decimal representing the percentage of tolerance acceptible}
}
\value{
The value return is a list of two vectors:
index | vector contained indexed values of critical points
values | vector containing the actual values
}
\description{
A very fast way to reduce the number of points in a time-series dataset. This algorithm takes a tolerance
range and will return all of the critical points/indexes over that range
}
\section{Functions}{
\itemize{
\item \code{}: 
}}
\examples{
point_reduce(1001:1100, .1)

#takes a sin curve of 99001 in length and reduces it down to 3008 points
point_reduce(sin(seq(1,100,.001)),.01)
}

