% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prep_ts}
\alias{prep_ts}
\title{Create a group of times endpoints for all groups to merge into existing dataset}
\usage{
prep_ts(dt, times, time = "time", group_by = NULL)
}
\arguments{
\item{dt}{A data.frame/data.table}

\item{times}{A vector of times to be added}

\item{time}{The name or index of the time column}

\item{group_by}{An index or string of the columns that serve as groups}
}
\description{
This function takes a set of times and a data.frame/data.table, and returns a data.frame that creates a time
for every unique group specified by `group_by` from `dt`
}

