% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reducePoints}
\alias{reducePoints}
\title{Reduce the number of points needed to plot time series data}
\usage{
reducePoints(x, tolerance)
}
\arguments{
\item{x}{A numeric vector}

\item{tolerance}{A decimal representing the percentage of tolerance acceptible}
}
\value{
The value return is a list of two vectors:
index | vector contained indexed values of critical points
values | vector containing the actual values
}
\description{
A very fast way to reduce the number of points in a time-series dataset. This algorithm takes a tolerance
range and will return all of the critical points/indexes over that range
}
\examples{
reducePoints(1001:1100, .1)

#takes a sin curve of 99001 in length and reduces it down to 3008 points
reducePoints(sin(seq(1,100,.001)),.01)
}

