% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_i}
\alias{str_i}
\title{Shorthand for getting indexes of a longer string}
\usage{
str_i(str_lng, str_shrt)
}
\arguments{
\item{str_lng}{A Character Vector}

\item{str_shrt}{A Character Vector}
}
\description{
Shorthand function for getting indexes of a shorter string on a longer string. Can use to get indexes
of the names of a data.frame/data.table
}
\examples{
str_i(letters, c('f','g','y'))
}

