% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dtdeduplicate}
\alias{dtdeduplicate}
\alias{vdeduplicate}
\title{De-duplicates a numeric vector/data.table/data.frame}
\usage{
dtdeduplicate(x, key)

vdeduplicate(x, returnIndex = FALSE)
}
\arguments{
\item{x}{A vector/data.table/data.frame}

\item{key}{If using a data.table/data.frame, the index for performing the operation on}

\item{returnIndex}{If TRUE, will return indexes, otherwise will return actual values}
}
\value{
Returns a vector/data.table/data.frame that has been de-deduplicated.
De-duplicated means that consecutive duplicate values are reduced to the first occurance.
}
\description{
These set of functions take a vector, or a datatable/dataframe (with a particular column)
and compress it by de-duplicating its values. Ideally, this is a good way to reduce the amount of data
for time-series use cases
}
\section{Functions}{
\itemize{
\item \code{}: 
}}
\examples{
vec <- c(1,1,2,2,3,3,4,4,5,5,4,4,3,3,2,2,1,1)
vdeduplicate(vec)
vdeduplicate(vec, TRUE)
}

