% Generated by roxygen2 (4.0.2): do not edit by hand
\name{voffset}
\alias{voffset}
\title{Retern a vector that is offset by one}
\usage{
voffset(x, offset = 1, end_value = "Last")
}
\arguments{
\item{x}{A vector}

\item{offset}{An integer}

\item{end_value}{The last value to add on to the vector, since it will be one short of the original `x`. NULL to not add anything on}
}
\description{
This function takes a vector x from 1 to N , and returns it from 2 to N, and N+1. This mainly used
for short hand purposes in other functions
}
\examples{
tmp <- 1:10
voffset(tmp)
voffset(tmp, end_value = 0)
}

