% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vreduce}
\alias{vreduce}
\title{Reduce the number of points needed to plot time series data}
\usage{
vreduce(vec, tolerance = 0.01)
}
\arguments{
\item{vec}{A numeric vector}

\item{tolerance}{A decimal representing the percentage of tolerance acceptible}
}
\value{
The value returned is a vector consisting of the indexed values that should be kept
}
\description{
A very fast way to reduce the number of points in a time-series dataset. This algorithm takes a tolerance
range and will return all of the critical points/indexes over that range
}
\examples{
vreduce(1001:1100)
vreduce(1001:1100, .1)

#takes a sin curve of 10000 in length and reduces it down to 2447 points
vreduce(sin(seq(1,100,length.out = 10000)))
}

