% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prep_ts}
\alias{prep_ts}
\title{Create a group of new_times endpoints for all groups to merge into existing dataset}
\usage{
prep_ts(dt, new_times, time_col = "time", group_by = NULL)
}
\arguments{
\item{dt}{A data.frame/data.table}

\item{new_times}{A vector of new_times to be added}

\item{time_col}{The name or index of the time_col column}

\item{group_by}{An index or string of the columns that serve as groups}
}
\description{
This function takes a set of new_times and a data.frame/data.table, and returns a data.frame that creates a time
for every unique group specified by `group_by` from `dt`
}
\examples{
dt <- data.table("times" = seq(1,5,length.out = 10), value = letters[1:10], grouper1 = c(rep("a",5), rep("b",5)))
prep_ts(dt,c(3.2,4.4),"times","grouper1")
}

